Remove battle ship class in favor of just game class
