+++ battle_ship/game.py
@@ -1,5 +1,5 @@
 from battle_ship.ship import Ship
-from battle_ship.player import Player
+from battle_ship.player import get_player, Player
 
 
 class Game:
@@ -12,3 +12,14 @@ class Game:
 
     def turn(self):
         pass
+
+
+def get_battle_ship_game():
+
+    player_2 = get_player("Russia")
+    player_1 = get_player("America")
+
+    return Game(
+        active=player_1,
+        inactive=player_2
+    )
